+++ client/src/app/appointments/appointments.component.html
@@ -5,6 +5,7 @@
       <a href="#" class="btn btn-info btn-lg">
         <span class="glyphicon glyphicon-home"></span> Home
       </a>
+      <button (click)="this.router.navigateByUrl('/aud-login')" type="button" class="btn btn-primary btn-sm">Audiologist Log In</button>
     </div>
 
     <!--<div class="aud-login" id="aud-button" style="float: right">
