+++ client/src/app/common/base-url.ts
@@ -1,3 +1,3 @@
-const baseUrl: string = 'REPLACE ME';
+const baseUrl: string = 'http://73.11.92.198:3333/';
 
 export default baseUrl;
