+++ client/src/app/common/clear-data-modal/clear-data-modal.component.css
