+++ client/src/app/common/clear-data-modal/clear-data-modal.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'clear-data-modal',
+  templateUrl: './clear-data-modal.component.html',
+  styleUrls: ['./clear-data-modal.component.css']
+})
+export class ClearDataModalComponent implements OnInit {
+
+  constructor() { }
+
+  public ngOnInit() {
+  }
+
+}
