--- client/src/app/common/generic-clear-memory-modal/generic-clear-memory-modal.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-generic-clear-memory-modal',
-  templateUrl: './generic-clear-memory-modal.component.html',
-  styleUrls: ['./generic-clear-memory-modal.component.css']
-})
-export class GenericClearMemoryModalComponent implements OnInit {
-
-  constructor() { }
-
-  public ngOnInit() {
-  }
-
-}
