+++ client/src/app/relogin-popup/relogin-popup.component.ts
@@ -1,9 +1,8 @@
-import { Component, OnInit, Inject } from '@angular/core';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
+import { Component, OnInit } from '@angular/core';
+import { MatDialogRef } from '@angular/material';
 import { Router } from '@angular/router';
-import { Utilities } from '../common/utlilities';
 import { ReloginPopupService } from './relogin-popup.service';
-import GenericClearMemory from '../common/generic-clear-memory';
+import ClearData from '../common/clear-data';
 
 export interface ReloginData {
   username: string;
@@ -17,24 +16,23 @@ export interface ReloginData {
 })
 export class ReloginPopupComponent implements OnInit {
   private loginFail: boolean = false;
-  private data: ReloginData = {username: '', password: ''};
+  private data: ReloginData = { username: '', password: '' };
 
   constructor(
     private diaglogRef: MatDialogRef<ReloginPopupComponent>,
     private router: Router,
-    private service: ReloginPopupService, 
-    private clearMemory: GenericClearMemory
-    ) { }
+    private service: ReloginPopupService,
+    private clearMemory: ClearData
+  ) { }
 
   public ngOnInit() {
   }
 
   private onLogout(): void {
     // clear all patient data in memory
-    this.clearMemory.clearMemory(false, () => {
-      this.router.navigateByUrl('/home');
-      this.diaglogRef.close();
-    });
+    this.clearMemory.clearData();
+    this.router.navigateByUrl('/home');
+    this.diaglogRef.close();
   }
 
   private onLogin(): void {
